moving files and some design improvements
